+++ src/cmdliner_manpage.ml
@@ -28,11 +28,11 @@ let s_arguments = "ARGUMENTS"
 let s_options = "OPTIONS"
 let s_common_options = "COMMON OPTIONS"
 let s_exit_status = "EXIT STATUS"
-let s_exit_status_intro = `P "$(iname) exits with:"
+let s_exit_status_intro = `P "$(cmd) exits with:"
 
 let s_environment = "ENVIRONMENT"
 let s_environment_intro =
-  `P "These environment variables affect the execution of $(iname):"
+  `P "These environment variables affect the execution of $(cmd):"
 
 let s_files = "FILES"
 let s_examples = "EXAMPLES"